fix conversion warnings when compiling c++
git-svn-id: https://dev.csoft.net/agar/trunk@9272 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ vg/vg_view.h
@@ -133,7 +133,7 @@ VG_ApplyConstraints(VG_View *vv, VG_Vector *pos)
 	switch (vv->snap_mode) {
 	case VG_GRID:
 		grid = &vv->grid[0];
-		ival2 = grid->ival>>1;
+		ival2 = (float)(grid->ival>>1);
 
 		r = VG_Mod(pos->x, grid->ival);
 		pos->x -= r;
